requirements.txt
numpy==1.23.1
pandas==1.5.3
Pillow==9.5.0
scikit_learn==1.2.1
imbalanced_learn==0.10.1
streamlit==1.24.0